add simple_calendar and tailwindcss-rails
i used the simple calendar gem since displaying a calendar is
something that is simple and shouldn't be reinvented. the gem
includes support for events (that may or may not be used in the
future) but the featureset seems to be exactly what i need for now.

i've also decided to use tailwindcss instead of other css processors
as that seems to be what the rails team prefer. there's also less
of a hassle since i won't have to add node dependencies unlike
with the other css processors (from what i've read at least)

since there's a css processor now, it looks like rails had added the
`bin/dev` stub which automatically runs foreman for us.