extract checklists from note body
when a line starts with some sigils (in this case the todo sigil `>`) it
is extracted and placed on the checklist area

there's also a quick summary page of all the month's checklists so
there's a good overview of the todos for the month. still trying things
out to see what works best.